requirements.txt
paramiko
python-dotenv
sh
pyyaml
